requirements.txt
asgiref==3.2.7
attrs==19.3.0
certifi~=2020.6.20
chardet==3.0.4
Django==3.0.5
django-cors-headers==3.5.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
idna~=2.10
importlib-metadata==1.6.0
more-itertools==8.2.0
packaging~=20.4
pluggy==0.13.1
py==1.8.1
PyJWT==1.7.1
pyparsing==2.4.7
pytest~=5.3.2
pytest-django==3.9.0
pytz~=2020.1
requests~=2.24.0
six~=1.15.0
sqlparse==0.3.1
urllib3~=1.25.10
wcwidth==0.1.9
zipp==3.1.0

colorama~=0.4.3
cryptography~=2.9.2
pip~=19.2.3
numpy~=1.19.2
pandas~=1.1.2
atomicwrites~=1.4.0
python-dateutil~=2.8.1
Pillow~=7.2.0
setuptools~=41.2.0
django-filter~=2.3.0

gunicorn==20.0.4
psycopg2-binary==2.8.5
python-dotenv